qt: bring back WA_PaintOnScreen for Win32
The doc says this only affects X11 but digging in the code it seems that on
Windows if paintEngine() is overridden and returning NULL this flag should be
used. That's what we had so far.

This was removed by 08f886b6e88f46ebeb1cfb2bc838c6fec8c81dfc

Also use positif #if testing so it's more readable what case does what.