qt: improve search logic
this includes a fix for the following flaw: if for any reason a
set_subcategory() call followed by no options was ever placed at the end
of the core option set, then this would have dereferenced memory beyond
the end of the set. (Note that such empty entries currently genuinely
exist in order to ensure that respective tree nodes are created for
modules to attach to, which will be fixed later).