qt: remove general tooltop on module checkbox group
With module selection controls which feature a groupbox of module
checkboxes, if individual modules have their own longtext, that is
displayed in a tooltip when hovering on them; for all others, the help
text for the entire control (groupbox) gets displayed, which is just
confusing from a UX point of view.

This removes the general groupbox tooltip, thus fixing this.

It remains for the textbox portion of the control however, so hovering over
that still provides the helptext!

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>