cmdline: reject short option chars with special getopt meaning
getopt can return ':' and '?' chars to indicate special conditions. there
is nothing to prevent such characters being assigned to options, and thus
there is a risk of misinterpretation, for instance misinterpreting a sign
of a missing option value error condition as a match for such a short
option. this change ensures that in the hypothetical event someone is
stupid enough to try to assign one of these chars to an option, we prevent
situations of misinterpretation of error conditions, at the acceptable
expense of that short option not being allowed to work at all.