cmdline: clarify missing option value error
for options that need an accompanying "value", users can either supply it
within the same command-line argument as the option itself, like
`--foo=bar`, or within the next argument, like `--foo bar`. (for short
options this would be `-fbar` and `-f bar` respectively).

it can be argued that it is possibly confusing to users to thus speak of
a missing "argument" when such an option "value" is missing. let's speak
of a missing option "value" instead.