avcodec: use a choice list for --avcodec-skip-idct as per description
gives much better UX in gui prefs.

the range was dropped, even though ranges are enforced (mostly) unlike
value-lists within the option system, and so it can be sensible to have
them in addition to value-lists, since the avcodec module has code handling
the value of this option which maps each valid value to a specific constant
and uses a default constant for unmapped values, so keeping the range would
be pointless here.