qt: setup filter variables on the main vout
Since the vout is now persistent, there is no need to store the variables on an
intermediate vlc_object_t (that was wrongly set to the interface during the
last UI merge).

IMPORTANT: the video filter settings will now only change for the main vout. We
need to define a way to change vout settings for other windows.