vout:win32: remove the display dimensions callback
The dimension to use is always set in area->vdcfg.display. The hwnd size always
matches this size since we resize it accordingly for each
DISPLAY_CHANGE_DISPLAY_SIZE call.

There is a slight delay between the moment the window tells the core of the
size change and the moment it's effective in the display. So resizing quickly
is not as smooth as when reading the HWND size directly.