vout:win32: add a flag to allow display modules handle the full display size
Right now all Windows display module render on the hvideownd which is already
cropped and positioned according to the PlacePicture results.
In external rendering we do not have this mechanism and so we need to be able
to render on the full display but cropping according to the PlacePicture result.

Ultimately that should be the only mode. D3D and OpenGL should have no problem
cropping and stretching in the proper area. WinGDI should also be able to only
draw on a portion of a HWND. It can already stretch.