vout: fix double mutex destroy
And move window deletion from vout_Close() to vout_Release(). That way, the
window is always valid during the lifetime of the vout. This simplify window
NULL checks.

Note: vout_IntfDeinit() is called from vout_Close(). So, theoretically, no
functions touching the window can be called after this call. This may change in
the future if we want to control the vout from functions directly.