modules: add type strictness to cat/subcat types
with cats and subcats both being plain integers it's very easy to
make a mistake mixing them up, storing one in a variable for another,
or passing the wrong variable to the wrong function.

switching to enums adds type strictness to help catch such mistakes.

in switching to enums there is of course additional opportunity to
exploit sequential numbering to achieve certain optimisations that
can be gained from this, but that it not done here. here we're simply
interested in catching mistakes through type strictness.

note that we simply retain the same cat/subcat numbering as before
here.