config: split the cat/subcat table and name/help getters
...to prepare for divergence in the data held for categories vs.
subcategories.

we replace here the generic helpers:
 - `config_CategoryNameGet()`
 - `config_CategoryHelpGet()`

with:
 - `vlc_config_cat_GetName()`
 - `vlc_config_cat_GetHelp()`
 - `vlc_config_subcat_GetName()`
 - `vlc_config_subcat_GetHelp()`

the new names conform to the format requested for the
`vlc_config_subcat_IsGeneral()` function added in MR 313.

opportunity was taken to remove the old style prefixes from the attributes
of `config_category_t`, affecting a few additional lines, avoiding
introducing inconsistency or new instances of the old style with the
addition of `config_subcategory_t`.