qt: create tree nodes where missing for plugin nodes
the advanced preferences tree construction code is designed to build up
the base cat/subcat tree from the core option set only, then attach
nodes for plugin sets to that tree.

the author(s) of that code made the mistake of assuming that the core
would always make use of all subcats, and thus that there would never be
a problem when it came to attaching plugin nodes. this assumption has
proved wrong, requiring redundant subcategory entries to be added to the
core option set as a temporary fix for placement issues
(1ca5bfaa1901551618ccac5fa97ac40babde7ab3).

5c18cff8d8cdba5784ca4bdb641fb483499117ed is a good example of the
fragility of this design. it removed the video-splitter entry from the
core set, which thus had the unintended and unnoticed consequence of
plugins targetting this category suddenly being incorrectly attached
to the parent cat node rather than appearing under the video-splitter
subcat child node as before.

having to keep 'unused' subcategories within the core option set just
to support proper plugin node placement, besides being fragile, also
has the disadvantage that if no plugins attach to it, we just end up
with useless empty nodes in the tree. it is certainly not a good long
term solution.

this patch addresses the problems just described by implementing the
ability for the plugin node handling portion of the tree contruction
code to add missing tree nodes if and when it needs to.

(this now makes 1ca5bfaa1901551618ccac5fa97ac40babde7ab3 redundant).