qt: improve efficiency of finding adv prefs tree nodes
fixed arrays of pointers for tracking the handful of cat and subcat node
items (but not plugin nodes), gives a much more efficient alternative to
the current solution of searching the tree, each time we need to attach
a new plugin node to the tree when building it.

this also proves to have a secondary benefit per the next commit.