config: get cat names from their general subcats
except in one pointless case, category names are identical to the names of
their 'general' subcats. if we fix that one case, as done here, we can
remove the name attributes from the category table and instead fetch the
name labels from the subcat table via general-subcat lookup, as done here.

i had to move the cat-name helper down to avoid a forward declaration. i
moved the cat-help helper down along with it for neatness.

the one case that had to be fixed for this to go ahead was the advanced
category. the tree node had the label "Advanced" while the panel title had
"Advanced settings". this inconsistency is unnecessary, especially since
the help text displayed underneath the panel title says "Advanced settings".