config: add general subcat data to cat data table
every category must have one general subcat, which is now made
more clear / enforced, by being captured in the category data
table.

the new helper that looks up the general subcat for a given
cat will be put to use by some later commits. for now it is
used only to reimplement the is-general helper to rely upon
the data table.