qt: simplify handling of general subcats
every cat has a general subcat, as defined in `vlc_config_cat.h`; let's
simplify things by merging the code that applies the general subcat
properties on top of the already established cat properties, rather
than doing this in a seperate step after creation of the category node.

note that the initial category item was given a `PrefsItemData` type of
`TYPE_CATEGORY`. once the general subcat properties were merged, it was
given the type `TYPE_CATSUBCAT`. this change now effectively wipes out
the need for this distinction. this will be tackled in a follow up
commit.

you may also notice that this makes more clear the fact that category
help text, as captured in the `vlc_config_cat.h` data table
`categories_array` goes entirely unused, being overridden by that of
the general subcat help text. removing this unused help text was a
point of contention when i previously tried to bring it up, so for now
it remains. (it is also unused in the macos gui).