config: sort category table into a preferred order
this is prep work for the next commit. the intention here is to have the
order of items in the category table reflect the preferred order for the
primary category nodes to be displayed within the advanced preferences
trees.

the current alphabetical ordering used in qt i feel is far from ideal, as
explain in the next commit.

the order i have chosen is based upon my perception of what categories
might be of the most interest to the most users:

 1. Playlist
 2. Interface
 3. Audio
 4. Video
 5. Input / Codecs
 6. Sout
 7. Advanced

i also re-ordered the subcat table, since it only involved moving one
block. this was purely for maintaining consistency.