qt: use preferred order for adv prefs tree cat nodes
up until now the entire tree has been simply sorted alphabetically. this is
not at all ideal because this means that the "advanced" category comes
first, when logically it should come last. this is particularly unfortunate
because upon switching to advanced/complete/"all" view, the panel shown is
one titled "advanced settings" with a warning underneath saying "use with
care". this category also happens to be a particularly messy one in terms
of its organisation. this mess and warning gives a bad impression and
potential scares off users.

so, let's allow the subcat and plugin nodes to remain alphabetical, but
let's sort the root category nodes into a preferred order. we can do this
simply by resorting to match the order of entries within the category
lookup table in `vlc_config_cat.h`, which was reorganised into a suitable
order in the previous commit.

note that the solution accounts for the possibility that not all cats will
be in use in the tree.