vout:win32: optimize the placement handling during window resizing
During fast resizing we might get a lot of WM_SIZE messages which are each
sending a CHANGE_DISPLAY_SIZE which may take time to get to the display thread.
We bypass a part of this lag by reading directly the current size of our parent
HWND (since we always place our HWND at 0,0 with the full size).

This is similar to what we have before when we handled all the HWND in this
part of the code. We read the window size directly on each size change.