macos: fix potential prefs tree option misplacement
when processing the core option set, if the subcat is a 'general' one and
we have a category node to attach the current option to, we attached it.
the alternate path was intended for otherwise attaching options of
non-general subcats to the relevant subcat node (when we have a node for
that subcat).

the problem with the implementation of this is what happens if the subcat
is considered a 'general' one, but should we not have a cat node. in this
case it would go down the incorrect path of trying to attach the option to
the last created subcat node, rather than performing the correct action of
ignoring it.