macos: fix placement of plugin node's targeting general subcats
as explained in the previous commit, plugins that reference 'general'
subcategories in their option sets are mishandled when it comes to creating
nodes in the plugin tree. the previous commit laid the groundwork for a
solution to this, and now here we actually fix the problem.

to summarise, if the target subcat is a general one, the plugin node is
supposed to be attached to the cat node, i.e. placed on the 2nd tier
alongside subcategory nodes for that category. instead it was mistakenly
attached to the previous `subCategoryItem`.

in fact if the first or only subcat specified in the plugin's option set is
a general one, `subCategoryItem` would be null, and thus the plugin node
object would just never be attached at all.