macos: fix multi-free unsoundness and duplicate plugin nodes in prefs tree
whilst most plugin descriptors make only a single `set_subcategory()` call
(indicating relevant location in the preferences tree for their options),
some make multiple calls, generating multiple 'subcat' entries within their
option sets.

it should be noted that the qt prefs tree construction code (currently)
handles this by essentially ignoring all but the first one; whether or not
this is the best design, vs. breaking up such option sets across the tree,
is a subject that is tackled in a piece of work to be submitted for review
later.

in any case, the macos prefs tree construction code here looked highly
suspect to me with regards to how it handles such situations. (i'm basing
this on reading code only, i have no Apple device available).

the loop processing each plugin:

 1. creates a `PluginBranch` item. this holds the copy of the config set
    returned by the `module_config_get()` call, which its destructor will
    `free()`.
 2. as the plugin's config set is iterated over, cat and subcat nodes are
    created as necessary (if nodes for them have not already been created)
    in direct response to any subcat entries (to ensure the existence of
    the parent nodes that the plugin node will be attached to in the tree).
 3. for other item types in the set (i.e. actual options along with
    'section' items), they are 'attached' to the `PluginBranch` item (added
    to it's `options` array). additionally the `PluginBranch` item itself
    is 'attached' to the relevant cat/subcat node (added to that node's
    `children` array) if it is not already present there.

so when it comes to plugin option sets that contain multiple subcat items,
this logic means that the `PluginBranch` object will get 'attached' once to
each and every unique subcat in the set (well, those that are directly
followed by at least one non-subcat item).

this fact alone is a problem, since we surely don't want copies of a
plugin's entire option set appearing in multiple places in the tree. it
could be a source of confusion, and creates a potential problem (which i've
not properly examined) when it comes to saving changes.

however it is worse than just an organisational problem - this duplication
of `PluginBranch` instances within the tree actually creates multi-free
issues when it comes to the matter of object destruction. when 'attaching'
`PluginBranch` objects to a cat/subcat we are adding its pointer to an
`NSMutableArray`, which apparently automatically calls `release` (the obj-C
equivalent of C++'s `delete`) on the items it contains when it itself is
destroyed. effectively, by placing the pointer of an allocated object into
such an array, we hand over ownership. the problem is that we are assigning
ownership of the same object to multiple arrays in these multi-subcat
cases. this means that these objects get released multiple times, and
furthermore via their destructors free their config sets multiple times.
ouch!

this commit fixes the problem very simply by keeping track of whether or
not we have already attached the plugin node somewhere, and only attaches
it to a cat/subcat if not. this prevents the issues described here, and
brings the handling of such multi-subcat option sets more into line with
the current qt handling.

an issue remains of potentially unnecessary cat/subcat node creation, which
will be tackled in a later commit.