macos: merge prefs tree node types
nodes for plugins are created in the preferences tree depending upon use of
'subcategory' entries within plugin option sets. while most plugins specify
'non-general' subcategories, and thus will appear on the 3rd tier of the
tree (cat -> subcat -> plugin), some specify 'general' subcategories (which
represent option sets shown when category nodes are selected), and in this
case the plugin node must appear on the 2nd tier alongside subcategory
nodes.

the existing code here did not support this however, with such nodes
instead mistakenly being attached to the previous non-general subcat node,
if available, which is completely wrong.

unfortunately implementation of a simple fix for this is blocked because
different object types are used to represent each of the three different
node types (category; subcategory; plugin) and everything is naively
designed upon the false expectation that the three tiers of the tree can
map purely to each of these respective object types, incompatible with
allowing for a mix of subcat and plugin nodes on the 2nd tier.

a significant overhaul of the node design is thus first required before we
can fix the actual misplacement issue. here i have taken the obvious
approach of unifying the three branch types into one, with an enum based
property identifying subtypes. yes, this requires a little extra memory,
but it's simple and gives us the needed flexibility.

the misplacement issue itself will be fixed in the next commit.

note, a few additional lines where changed in taking the opportunity to
rename `subCategoryItem` to `subcategoryItem`.