macosx: respect hidden cat/subcat "hack"
Certain sets of options in the core option set use a cat/subcat of -1 as
a "hack" to avoid those options showing up in the GUI prefs set. This is
the case for options like --help for instance. While the Qt GUI plugin
respects this, the MacOSX one was missing checks for this, and so surely
(I don't have a mac so I can't look and see) it was including them when
it shouldn't.

(cherry picked from commit 05e2f75ad01decaf1c83c7f45e6cabbf3833d5b1)