contrib: qt: Add static plugins/qml modules to the static pkg-config
Get the library/pathes from the installed PRL files in PREFIX/plugins or
PREFIX/qml and add it to Libs.Private in the .pc files. So the libraries are
kept in those folders rather than copied in the libs directory.

We get exactly the libs and order that qmake would use when linking.