contrib: fxc2: Fix building fxc2 while targeting non x86 architectures
As fxc2 relies on a precompiled binary for x86, which we need to run
at build time, we need to look for an x86 cross compiler, in case the
current target for contribs isn't x86.

Remove the patch for adding a meson build system to fxc2. fxc2 consists
of one single source file, and when building with meson, one must
create a cross file in order to specify using a different cross compiler,
just passing the name as a parameter isn't possible. Therefore, just
resort to manually compiling the single file, instead of adding a
new buildsystem that upstream doesn't contain and which makes cross
compiling with arbitrary compilers harder.