resource: respect vout order
The first vout created (the main one) will now always be the first vout of the
array. It's important to keep the same reference on the first vout, since its
window plugin is likely linked to an interface. This commit ensure that this
first vout will always be the first vout used by the first video decoder or
visualiser.