qt: merge lazy creation of related widgets
the search filter for the advanced preferences tree should always be
lazily created at the same time we lazily create the tree widget itself.
there would never be a situation where this function would find one already
created and not the other. we can thus merge the condition blocks here.