qt: reorganise prefs construction code
it makes it much easier to understand the makeup and contruction of
each individual view and of the entire window this way, with the
components and construction code for each view grouped more
logically together. even more so when a new view is added in later
work of mine. grouping by type of component and interweaving the
creation of the views just make things unnecessarily harder to
follow and work with.