declink: fix issue with option choice labels
the integer mapping should not appear like this in these labels. it just
creates a mess. it is the job of the CLI interface to show the values
besides this text such as to give users this mapping. in GUIs showing
users this mapping is completely unwanted.

these now happen to be identical to those in sdiout and thus consistent
and would reduce translator burden if translated in future.