qt: split out password control creation into dedicated type
the string type was unusual in having a special extra boolean parameter to
select between an ordinary string type control and a password type one,
rather than having a password specific type which derived from it, as
everything else does.

splitting out password controls into their own derived type simplifies
things a little.

it also achieves two other things:
 - it puts us into a position from which we can further customise password
   controls, as i have done in some later work to be submitted after this.
 - it makes it easier/cleaner to add further string derivative controls,
   as i also have done in at least one instance in other work.

note that we are also removing a boolean param from the constructors of
`IntegerListConfigControl`. this type only had this param because the
string type did, in connection to the use of generic macros in the simple
preferences code.