contrib: gnutls: Remove one unnecessary level of ifdefs
The aarch64 assembly in gnutls still needs to be disabled on windows
in general though; the issue is that it unconditionally uses ELF
specific directives (.type, .size, .section), so the disabling isn't
related to any specific issue in clang, but only that gnutls' assembly
for this architecture isn't portable.

Add a comment to clarify the issue.