playlist: temporize on errors
When repeat/loop is enabled, and items fail to play, VLC enters an
infinite liveloop and floods errors.

We cannot just stop on errors, because retrying indefinitely may be the
expected behavior:
<https://trac.videolan.org/vlc/ticket/18582#comment:3>

Instead, wait some delay before starting the next item, depending on the
number of consecutive errors:
 - 1st error: 100ms
 - 2nd error: 200ms
 - 3rd error: 400ms
 - 4th error: 800ms
 - 5th error: 1.6s
 - further errors: 3.2s

A single successful playback resets the errors counter.

Note-from-committer:
--------------------
This is not a complete and correct fix, but this hides the issue for
our users for the vast majority of the cases.
Once a proper fix is done, please revert this.

Ref #5901

Fixes #6245 #6339 #7305 #7798 #8893 #9230 #11066 #11961 #12104
      #12909 #13878 #14679 #15805 #16697 #17660 #18582 #18942 #19284
      #19313 #20365 #21564 #21672 #21897 #22118