qt: fix double translation of subcats in advanced prefs
the config_CategoryNameGet() and config_CategoryHelpGet() functions already
perform translation, so the use of qtr() instead of qfu() on the result of
calling those functions is an undesirable double translation.

this issue, combined with two others, results in a null dereference crash
(simply opening the advanced preferences view in the Qt GUI) under
conditions that arose in some work i was doing. those other issues
specifically are:
 1) the lookup table used by those functions is incomplete, missing
    records for a few subcats, thus a null is returned when looking them
    up. (this is addressed in a follow up commit).
 2) the vlc gettext wrappers crash when given a null pointer. (the subject
    of a separate patch discussion).

the explicit null check for help text has been removed here, since every
subcat should have help text along with a name in the lookup tables.

Signed-off-by: Pierre Lamot <pierre@videolabs.io>