qt,macos: (hotkeys) robustify hotkey item identification
the existing checks:
 - assumed that any core config item (excluding special class `0`
   hint/cat/section types) that had a name starting with "key-" was a
   non-global hotkey item.
 - global hotkey item identification is similar, except looking for a
   prefix of "global-key" (Qt only).

there is no guarantee that a non-hotkey option that begins with "key-"
will never be introduced. similarly, there is no guarantee that a
hotkey option will never be introduced with a different prefix. if
either of those hypothetical situations where to occur, the hotkey
control would mishandle them.

here we:
 - check whether or not the item type matches `CONFIG_ITEM_KEY` to
   properly identify hotkey options.
 - make use of the "global-" prefix prepended to the global option
   name by the hotkey creation macro, to make a more reliable
   distinction between global and non-global.

obviously one imperfection would be that if there ever was a hotkey
option introduced with a name starting with "global-" then this would
be misidentified, but that is very unlikely, and the only way to
achieve complete perfection would be to introduce a global hotkey
subtype.