qt: (hotkeys) remove incorrect hotkey table label tooltip
the item pointer provided to the constructor is that of a single hotkey
config item. using the longtext from this item in a tooltip on the label
for the entire hotkey table makes zero sense, so let's remove it.

 - for the table shown in the simple view, the "key-play" option was
   arbitrarily used for this.
 - for the table shown in the advanced view, the text from the first
   hotkey config item encountered was used.

(cherry picked from commit b0385e9324a0f342e51bf5b61a50904adb4d47dd)