qt: (hotkeys) remove useless `KeySelectorControl` constructor param
this widget is that through which all hotkeys are edited.

as just explained in the previous commit, the item pointer provided to
the constructor is always that of a single hotkey config item.

the point in passing along an item pointer was that the base control class
requires one. note that all other controls target a single config item and
so it makes sense for them, but does not make any sense for this. for
whatever reason, someone thought it made more sense to pass along an
arbitrary hotkey item pointer rather than null (possibly in relation to
the fact of displaying the incorrect tooltip removed in the previous
commit).

since the param is entirely useless for this widget, this removes it and
passes along null to the base control constructor instead. this control
simply does not need to be linked to an individual item.

note that the line removed involving "key-play" is deliberate, that was
the arbirarily chosen item to use for the widget in the simple view. (the
`p_config` var is still needed for other things though).