qt: (hotkeys) rework conflict check
this replaces fetching a list of conflicts to react to, with
iterating over the table and looking for conflicts as we do so.

the primary purpose of this is not efficiency, though it is more
efficient; it is actually to support the 'multi-key assignment conflict
check' fix to come in a subsequent commit in this set.

furthermore it is needed to allow improvement of the conflict checker
regarding avoiding translation (we need to be able to compare the custom
data attributes for that, not the displayed text, as `findItems()` does).