qt: (hotkeys) simplify conflict checking wrt. "unset"
firstly, note that text displayed in the table is (supposed to be)
translated (it is not currently it seems for the global column), whilst
each key mapping cell holds a data attribute containing an untranslated
copy of the actual mapping value text.

the logic here:
 1. asked `VLCKeyToString( i_vlckey, true )` to convert the captured
    key/key-combination integer to string form (with translation).
 2. searched the display text of the relevant table column to find entries
    that matched.
 3. looking only at the first match (conflict), if there is one, adjusted
    the dialog for displaying the conflict condition.

if conversion failure occurs in step 1, the conversion function returns
"Unset". the behaviour then is such that we allow storing this text for
the item, thanks to step 3 ignoring conflicts with entries that have an
unset state (whether "Unset" or empty).

the change made here concerns the complexity and inefficiency of how the
conflict check is skipped for such a situation, and the legitimacy of the
is-empty side of it.

so firstly, wrt. the former: if the data value of the inspected conflict
item is "Unset", then this must be because the corresponding display text
was "Unset" (the translated form), and thus that the search text (the
converted key) was same. (we can safely ignore the hypothetical case of
translation resulting in the same text for both "Unset" and some actual
key name as being unrealistic, and that will be of no concern when we
switch to using the non-translated text for conflict checking in a later
commit in this set anyway). given this, we can simplify and optimise
handling this situation by checking if step 1 returned "Unset" up front,
and directly accepting it if so. this is a much more simple check, and
saves effort. (note that we no longer check "Unset" against the "existing
keys" set, but we would never expect a legit conflict there anyway).

secondly, the situation of the data value being empty i believe is bogus.
an item's data value should only be empty where it is in a "proper" unset
state (not the "Unset" text situation just mentioned), which can only occur
either through the saved value being so; from use of the `del` key; or from
using the `unset` button in the edit dialog. in all three cases the display
text should also be empty. such an item could thus only be in the conflict
search results if the search text itself had been an empty string. since an
empty string does not represent any key/combination, and since the int to
string conversion always gives "Unset" upon failure, it should not be
possible for an empty string search to take place, and thus we would never
encounter a conflict item with an empty data value. we can thus ditch the
check for that.